add class to MIR script
